% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-template.R
\name{create_shiny_project}
\alias{create_shiny_project}
\title{Create a project using Shiny app template}
\usage{
create_shiny_project(path)
}
\arguments{
\item{path}{}
}
\description{
Create a project using Shiny app template
}
\examples{
NULL
}
