% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{filter_active}
\alias{filter_active}
\title{check if a filter is active in a teal module}
\usage{
filter_active(datasets)
}
\arguments{
\item{datasets}{instance of teal filtered datasets class}
}
\value{
boolean, TRUE if a filter is applied, FALSE otherwise
}
\description{
check if a filter is active in a teal module
}
