% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-theme.R
\name{mkshinytheme}
\alias{mkshinytheme}
\title{Return the URL for a Shiny theme}
\usage{
mkshinytheme(theme = "teal")
}
\arguments{
\item{theme}{Name of the theme.}
}
\description{
The result of this function should be used as the \code{theme} argument for
\code{\link[shiny]{bootstrapPage}}, \code{\link[shiny]{fluidPage}},
\code{\link[shiny]{navbarPage}}, or \code{\link[shiny]{fixedPage}}.
}
\examples{
library(shiny)
if (interactive()) {
  shinyApp(
    ui = fluidPage(
      theme = mkshinytheme("teal"),
      actionButton("btn", "Button", class = "btn btn-primary")
    ),
    server = function(input, output) { }
  )
}
}
