% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkshinythemes-package.R
\docType{package}
\name{mkshinythemes-package}
\alias{mkshinythemes}
\alias{mkshinythemes-package}
\title{mkshinythemes: Merck Branded Themes and App Templates for 'shiny' and 'shinydashboard'}
\description{
Merck themes and app templates for 'shiny' and 'shinydashboard'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stash.merck.com/projects/BARDS/repos/mkshinythemes}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
