% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_data_path}
\alias{set_data_path}
\title{update data file configuration setting}
\usage{
set_data_path(path)
}
\arguments{
\item{path}{path to directory containing data files
used within the Shiny application}
}
\value{
Used for side effects
}
\description{
update data file configuration setting
}
