% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tooltip_text}
\alias{tooltip_text}
\title{style a tooltip produced by the tippy package}
\usage{
tooltip_text(text, font_size = 16)
}
\arguments{
\item{text}{String for text in tooltip}

\item{font_size}{Font size (in pixels)}
}
\value{
HTML with font size applied
}
\description{
style a tooltip produced by the tippy package
}
