% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pilot2wrappers-package}
\alias{pilot2wrappers}
\alias{pilot2wrappers-package}
\title{pilot2wrappers: R Consortium R Submission Pilot 2}
\description{
Insert Package Description.
}
\author{
\strong{Maintainer}: xxx \email{xxx@xxx.com}

Other contributors:
\itemize{
  \item copyright [copyright holder]
}

}
\keyword{internal}
