% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{pilot2wrappers-package}
\alias{pilot2wrappers}
\alias{pilot2wrappers-package}
\title{pilot2wrappers: R Consortium R Submission Pilot 2}
\description{
Shiny application used within the R Consortium R Submissions Working Group Pilot submissions.
}
\author{
\strong{Maintainer}: Eric Nantz \email{theRcast@gmail.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Heng Wang
  \item Gregory Chen
  \item Eli Miller
  \item Ning Leng
}

Other contributors:
\itemize{
  \item copyright [copyright holder]
}

}
\keyword{internal}
